Рефакторинг компонента предложения
